pyproject.toml
[project]
name = "pydentity"
version = "0.1.0"
description = "A Python microservice for managing authentication and identity verification."
authors = [{ name = "AymanKastali" }]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlmodel>=0.0.22",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pyjwt[crypto]>=2.10.0",
    "argon2-cffi>=23.1.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pydentity"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.coverage.run]
data_file = ".pytest_cache/.coverage"

[tool.ruff]
target-version = "py314"
line-length = 88
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TC"]

[tool.ruff.lint.per-file-ignores]
"src/pydentity/infrastructure/web/routes/*.py" = ["B008", "TC001"]

[tool.ruff.lint.flake8-type-checking]
runtime-evaluated-base-classes = ["pydantic.BaseModel", "sqlmodel.SQLModel"]
runtime-evaluated-decorators = ["dataclasses.dataclass"]

[tool.mypy]
python_version = "3.14"
strict = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "debugpy>=1.8.0",
    "codespell>=2.3.0",
    "pre-commit>=3.7.0",
    "watchfiles>=1.1.1",
]
